flask-app/requirements.txt
scikit-learn
pandas
numpy
markdown
python-dotenv
google-generativeai
flask